more work for getting rid of jsession

+++ server/tomcat/src/mv/web/rest/Rest.java
@@ -13,6 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 
 
 import mv.core.util.DbCloser;
+import mv.core.util.ExternalException;
 import mv.core.util.Streams;
 import mv.db.DbFactory;
 import mv.db.RecordDb;
@@ -64,7 +65,10 @@ public class Rest extends HttpServlet {
 			String urlPath = request.getPathInfo();
 			String id = urlPath.substring(urlPath.lastIndexOf("/")+1);
 			db = DbFactory.instantiateRecordDb();
-			String user = db.getSession(request.getHeader("X-Session"));
+			
+			String user = db.getSessionUserId(request.getHeader("X-Session"));
+			if (user == null)
+				throw new ExternalException("Unknown session");
 			
 			String result = db.getObjectWithClassAndId(user, clazz, id);
 			System.out.println("response: " + result);
@@ -97,9 +101,12 @@ public class Rest extends HttpServlet {
 					System.out.println(kv.getKey() + ":" + v);
 			String json = Streams.readFullyString(request.getInputStream(), "UTF-8");
 			System.out.println("doPost: " + json);
-			String user = (String) request.getSession().getAttribute("user");
 			
 			db = DbFactory.instantiateRecordDb();
+			String user = db.getSessionUserId(request.getHeader("X-Session"));
+			if (user == null)
+				throw new ExternalException("Unknown session");
+
 			String results = db.putObjectWithClazz(user, clazz, json);
 			
 			response.setContentType("text/json; charset=UTF-8");
@@ -124,9 +131,12 @@ public class Rest extends HttpServlet {
 		try
 		{
 			System.out.println("doDelete" + request);
-			String user = (String) request.getSession().getAttribute("user");
 		
 			db = DbFactory.instantiateRecordDb();
+			String user = db.getSessionUserId(request.getHeader("X-Session"));
+			if (user == null)
+				throw new ExternalException("Unknown session");
+
 			String urlPath = request.getPathInfo();
 			String id = urlPath.substring(urlPath.lastIndexOf("/")+1);
 			db.deleteObjectWithClazz(user, clazz, id);
